exercise-1-fix/src/main/java/za/co/entelect/dojo/validation/CardValidationImpl.java
package za.co.entelect.dojo.validation;

public class CardValidationImpl implements CardValidation {

    @Override
    public boolean isValidCardNumber(String ccNumber) {
        int sum = 0;
        boolean alternate = false;
        for (int i = ccNumber.length() - 1; i >= 0; i--) {
            int n = Integer.parseInt(ccNumber.substring(i, i + 1));
            if (alternate) {
                if (n > 9) {
                    n = (n % 10) + 1;
                }
            }
            sum += n;
            alternate = !alternate;
        }
        return (sum % 10 == 0);
    }

    @Override
    public void isValidTrackData(String trackData) {
    }
}
